+++ src/point-free.js
@@ -1,0 +1,5 @@
+export { default as ap } from './point-free/ap.js';
+export { default as chain } from './point-free/chain.js';
+export { default as liftA2 } from './point-free/liftA2.js';
+export { default as map } from './point-free/map.js';
+export { default as pluck } from './point-free/pluck.js';