+++ src/point-free/liftA2.js
@@ -1,0 +1,5 @@
+import {curry} from 'ramda';
+import ap from "./ap.js";
+import map from './map.js';
+
+export default curry((f, ma, mb) => ap(map(f, ma), mb));