+++ src/promise/TypeRepresentative.js
@@ -1,0 +1,9 @@
+// see https://github.com/fantasyland/fantasy-land/tree/v4.0.1#type-representatives
+import of from './of.js';
+
+const
+	TypeRepresentative = {
+		"fantasy-land/of": of
+	};
+
+export default TypeRepresentative;