+++ src/promise/of.js
@@ -1,0 +1,6 @@
+
+const
+	// of :: a -> Promise e a
+	of = value => Promise.resolve(value);
+
+export default of;